Add top level error handler
`AppError` is rendered at the top level if an uncaught error occurs during rendering, in lifecycle methods,
and in constructors of any component in the whole react tree.

This is recommended to handle errors gracefully in React 16 because now the whole component tree
will be unmounted if an uncaught error happens.

This is implemented as a reusable `ErrorBoundary` component, this way it can be used in a more granular way
to handle errors in different parts of the app while the rest remain interactive.